+++ src/services/index.ts
@@ -35,3 +35,6 @@ export * from "./search.service";
 export * from "./checkout.service";
 export * from "./event.service";
 export * from "./admin.service";
+export * from "./bid.service";
+export * from "./contact.service";
+export * from "./media.service";